The case when you could edit binary data because of the ProtectBinary variable is now handled as well. In this case, the (maybe edited) value is kept, when no file was uploaded. I easified the IF/ELSE-Logic into those parts:
1. If the file is valid, always use the data of the file and
overwrite any existing data or any input made in the textarea, if
ProtectBinary is set to FALSE.

2. If the file is malicious, the previous data will not be overwritten. If
ProtectBinary is set to FALSE, the input of the textarea will be used
instead.

3. If the file is empty, the previous data will be erased.

Please double-check, even though it's okay for me and I checked all
edit/insert actions with both ProtectBinary set to blob and to false.