First batch of PHP4/MySQL compatibility/performance changes. Everything should be non-intrusive for installations meeting our new 4.1.0/3.23.32 specs.
I will commit my files every now and then for easier code review and to let you see which constructs I replace. Jump in if there are any objections. The basic thing I'm doing is documented in the changelog.

I'll continue with files in the libraries/ subdirectory and will move to the root-level files in last instance.