No need to ignore anything in this dir anymore
We don't use Pootle and advisory rules are generated directly.

We don't use Pootle and advisory rules are generated directly.

We don't use Pootle and advisory rules are generated directly.

We don't use Pootle and advisory rules are generated directly.

We don't use Pootle and advisory rules are generated directly.

We don't use Pootle and advisory rules are generated directly.

We don't use Pootle and advisory rules are generated directly.

We don't use Pootle and advisory rules are generated directly.

We don't use Pootle and advisory rules are generated directly.