Correct and complete license notices for bfShapeFiles.
I downloaded bfShapeFiles from http://bfshapefiles.sourceforge.net/
Exact URL was:
  http://softlayer-dal.dl.sourceforge.net/project/bfshapefiles/bfshapefiles/0.0.2/bfShapeFiles-0.0.2.zip
  and the sha256sum of that zip file was 40c3e575397cede5342493ca12950fb0e2e9ef6b1e7fd1bf1337e9ac9f24f11c

I see that the file in that zip file named
bfShapeFiles-0.0.2/lib/ShapeFile.lib.php appears to be nearly identical to
this ShapeFile.lib.php in the phpMyAdmin repository.

Thus, I have taken the copyright and license notice from
bfShapeFiles-0.0.2/README.TXT and included it herein, so that full copyright
notice is properly displayed.

Furthermore, I note that since the notice does not refer to a specific
version of the GNU GPL, and that bfShapeFiles-0.0.2/GPL.TXT includes a copy
of GPLv2.  This means that per GPLv2§9, the
bfShapeFiles-0.0.2/lib/ShapeFile.lib.php was licensed to the phpMyAdmin
project under GPLv2-or-later.  I have thusly updated the @license link to so
indicate.