Add DCO and note on how contributions are licensed
phpMyAdmin begins using a DCO as of 2013-12-01 to ensure changes are GPLv2-or-later.  Conservancy has drafted and approved the included Developer's Certificate of Origin ("DCO") and a related patch to the README to create a procedure whereby future contributors can use Signed-Off-By: in Git commits to confirm the patch is licensed under terms compatible with GPLv2-or-later.

Signed-Off-By: "Bradley M. Kuhn" <bkuhn@ebb.org>
Signed-Off-By: Tony Sebro <tony@sfconservancy.org>
Signed-Off-By: Denver Gingerich <denver@ossguy.com>