Drop flashcanvas
Let's drop flashcanvas, which we use for rendering charts with MSIE 8
and older. I've tested it a bit and in none of the uses it works
properly, so it really does not work as a drop in replacement for native
browser support and it would probably need some adjustment in our code.

Also it is problematic from licensing point of view as we do not ship
source for it js/canvg/flashcanvas.swf.

Signed-off-by: Michal Čihař <mcihar@suse.cz>