Fixes the problem with the action panel during row editing in chrome(Bug 4228)
This bug exists only in chrome and it is because of the way chrome renders
fieldsets which are used to display the action panel. Also the rows are
displayed in a random fashion unlike in firefox where they are sorted by the
primary key. Hence to fix this problem I surrounded each row wih a div tag with
float left property.

Signed-off-by: Aditya Sastry <ganeshaditya1@gmail.com>