Fixes the action panel problem when editing rows in chrome(Bug 4228)
The action panel which is a fieldset is rendered differently in chrome.
Also the rows are listed in a random fashion unlike in firefox where
they are sorted by the primary key. I fixed this problem by enclosing the
rows in a div tag with float left and clear both styles.

Signed-off-by: Aditya Sastry <ganeshaditya1@gmail.com>