Multicolumn sorting feature request(1507) Implementation part(1)
I divided the implementation for this feature request into three parts.
This is the part 1 of those three. My plan is not to break any existing
code for implementing this but rather build on top of it. To implement this
I grab the URL for single column sorting and modify it by adding to it
additional columns to sort by. Once I generate this URL I redirect my page
to this URL. This is the part 1.

In part 2, I will be modifying the existing code slightly to show visual cues
for select queries involving ordering by multiple columns.

In part 3 I will write any tests required.

Signed-off-by: Aditya Sastry <ganeshaditya1@gmail.com>