Remove workaround for older MySQL versions
Remove workaround originally implemented for MySQL < 5.0.2, but also
useful in 5.1.x times for listing tables within database when
permissions on one of database directories are wrong.

I've tested this with several MySQL versions and it does not seem to be
needed now, the INFORMATION SCHEMA works correctly even if some of the
database directories have wrong permissions.

Signed-off-by: Michal Čihař <michal@cihar.com>