Get rid of errors for slave servers
I had a bunch of errors saying that several variables don't exist when I clicked on a database, so I used the `global` keyword to get them to get their values correctly. They are defined in `libraries/replication.inc.php`, but since we are doing `include_once` they are not available to this function after the first iteration (correct me if I'm wrong). Also, in my case `$server_slave_Wild_Do_Table` and `$server_slave_Wild_Ignore_Table` were arrays with only one entry that was an empty string. This could be an error in my server connecting to the master server or vice versa, but this also produced an error when executing `PMA_extractDbOrTable` since `$list[1]` is undefined.