Implemented new functionality for voice commands!
This opens a new chapter for phpmyadmin.

1. You can now switch tabs using voice commands like Show Databases, Show Users.

Developer Suggestions needed to guide the development and future of
this feature.
This seems to me a crucial feature for a software like phpmyadmin
which is meant to make the user experience convenient when managing databases.

We could possibly add more features like Drop Database which would directly
lead us to a page where a list of databases would be listed
and we can choose which database to drop simply by naming it.

The possibilities are endless. And I am excited! :)

Signed-off-by: Pranav Sharma <vanarpsharma@gmail.com>