Added the new SSL check disable flag
MySQL 5.6 or later requests peer_name validation of the SSL.
This causes most self signed certificates to fail validation in mysqlnd.
In PMA, you only get a vague 2002 connection error as a result.
As of PHP 5.6.16, a new flag exists to disable the check and allows the
connection to proceed.
See https://bugs.php.net/bug.php?id=68344 and #11838

Signed-off-by: Joel Hutchinson joel.hutchinson@onlinecommercegroup.com