Remove session caching for MySQL charsets
This cache makes the session data quite huge giving performance penalty
when loading the cache. Loading the data again from MySQL server is
faster than this.

Also the charsets data is really needed only on few pages, the overall
impact of this caching is negative.

Caching the data per request however is useful.

Signed-off-by: Michal Čihař <michal@cihar.com>