Fix locking issues when importing SQL
When running import (or generally any SQL query), some tables might be
locked. This will cause queries to not locked tables to fail. And this
is exacly what we do with control user link.

So let's have separate link for control queries even if we're using same
credentials as for main link.

Fixes #12219

Signed-off-by: Michal Čihař <michal@cihar.com>