Remove token from GET requests
The CSRF token really should be used only in POST requests. The reason
for that is that it's a bit harder to get to if it is in request body
(with POST) compared to GET request, where it is in the URL (being
easily available in server logs).

Also this will make the URLs look nicer ;-).

This change will definite break some functionality, but since #6297 most
of the code should be safe and remaining bugs can be fixed for upcoming
release.

Signed-off-by: Michal Čihař <michal@cihar.com>