Simplified getting error from preg_match
It's enough to use error_get_last to obtain the message without need to
play with PHP configuration and error handling.

The only tricky thing is to know that there is new error. With PHP 7,
there is error_clear_last() which ensures this, on older PHP we need to
trigger *another* error to be able to detect that new error comes.

Signed-off-by: Michal Čihař <michal@cihar.com>