Handle AM/PM marks in date on our own
This is same as with months and days in week - we can not rely on the
PHP providing sane result. Also this way we can properly translate it to
all supported languages.

For example on OS X, the strftime function does not seem to respect PHP
locales and always use system ones.

Fixes #12913

Signed-off-by: Michal Čihař <michal@cihar.com>