Remove ServerLibraryDifference_DisableWarning setting
This is getting more and more problematic these days as most of Linux
distributions ship MariaDB libraries, while many people run MySQL
servers. As MariaDB claims the protocol compatibility should stay, we
should not warn in this case.

Of course there is option to add this exception to the warning (so far
the exception is there only other way - when using MySQL libraries to
connect to MariaDB server), but we would be ignoring most cases already
and I think complete removal is better than adding exceptions.

Overall I don't think this warning is doing any good these days.

Signed-off-by: Michal Čihař <michal@cihar.com>