Rewrite SSL verification code for curl
It seems that the curl extension is quite tricky what can be set. You
can not set both CURLOPT_CAINFO and CURLOPT_CAPATH. Unfortunately some
systems seem to have built in defaults for one of these which can not be
figured out of PHP. So let's just try the three options which make
sense.

Issue #12846

Signed-off-by: Michal Čihař <michal@cihar.com>