Update logo images
The old ones did not have proper transparency (due to lack of support
for that in some browsers by the time they were created).

As we don't seem to have sources for the original logo, I've used
slightly newer version which was used on t-shirts and stickers.

Fixes https://github.com/phpmyadmin/phpmyadmin/issues/13218

Signed-off-by: Michal Čihař <michal@cihar.com>