Rewrite export alias defining
The old code did load all colums for all tables in all databases in
export. This can easily explode for server wide exports or exports on
database with lot of tables.

It has been rewritten in a way to load required data on demand from
server. This will make aliases defining slightly slower, but I think
it's acceptable price for making the export work on larger databases.

Fixes #13008

Signed-off-by: Michal Čihař <michal@cihar.com>