refactor: cast the foreign field to an array prior to using it
right now, there is no way to select multiple columns for an index, even though mysql allows it. Since the destination_foreign_column will always be a string at the moment, I am casting it to an array prior to using it, as the code expects it to be one.

we were passing a string into the function to create sql to create a foreign key constraint. This function expects an array, since foreign keys can reference multiple columns.

Fixes Issue #13941

Signed-off-by: Leonid Sklyut <lsklyut@kbra.com>