Fix merge conflicts in QA_4_8 from this pull request
Signed-off-by: Isaac Bennetch <bennetch@gmail.com>

Fix bad callers to getError where null was being passed.

Fixes #14552: Considering how function signatures work in PHP, this should
be a reltatively 'safe' change, as the mysqli lib currently would still
do a check for cases where the passed in $link was either null or false.
Otherwise the bug's source appears to be a common mistake in PHP:
null != undefined

Signed-off-by: Andy Baugh <thomas.baugh@cpanel.net>