Fixed condition on the emptiness of enum type criteriaValues array (#14648)
The if condition has a call to the empty function on the criteriaValues array, but in the considered case that array only contains a '0', so this makes it look empty to the function. I added a control on the first element of the array, to avoid this specific situation.
It's my first contribution, so I hope I did well with the branches. Let me know if there's something missing or to do, I'm happy to help.

Signed-off-by: Alessandro Dolci <alessandro.dolci@pm.me>
(cherry picked from commit 3f5fff3f4d474fd9dd67f54ff2366d2a466e0c07)
Signed-off-by: Maurício Meneghini Fauth <mauriciofauth@gmail.com>