Bugfix 14631: Enum with comma produces incorrect search dropdown
When a comma is used withing the value of an ENUM field the search dropdown incorrectly splits it.

The bug occurs due a wrong split logic. The split function was assuming just a single quote as delimiter. This commit alters the delimiter to a single quote followed by an comma.

Signed-off-by: Wellington Braga <wjbraga@artit.com.br>