Resolved the issue #14298
This is in reference to resolving issue 14298.The problem was that when no query is passed and the user run it. The application shows 500 error.This was resolved by checking if the query is null or not. If it is empty then a message is displayed informing about the empty query

Signed-off-by: Gaurav Punjabi <gaurav.punjabi6@icloud.com>