Add check for empty input to change_collation
Currently, if an empty input is provided to change_collation, the behaviour is as if no user input were made.
This causes a bug in db_operations.php, where a blank form will be returned to the user instead of an error message.

By checking for empty($_REQUEST['db_collation']) separately, we can catch bad user input and return the appropriate error message.

Signed-off-by: Si Jie <sijie123@gmail.com>