Replace Array constructors with literal notation
Use of the Array constructor to construct a new array
is generally discouraged in favor of array literal
notation because of the single-argument pitfall and
because the Array global may be redefined. The exception
is when the Array constructor is used to intentionally
create sparse arrays of a specified size by giving
the constructor a single numeric argument.

Signed-off-by: Maurício Meneghini Fauth <mauricio@fauth.dev>