Use apply tag instead of filter and fix for loop
https://twig.symfony.com/doc/1.x/tags/apply.html
https://twig.symfony.com/doc/2.x/deprecated.html#tags
"The filter tag is deprecated in Twig 2.9. Use the apply tag instead (the Twig\TokenParser\FilterTokenParser classes is also deprecated)."
and "Adding an if condition on a for tag is deprecated in Twig 2.10. Use a filter filter or an "if" condition inside the "for" body instead (if your condition depends on a variable updated inside the loop)"
Signed-off-by: William Desportes <williamdes@wdes.fr>