Allow Windows TravisCI to fail + fixes
The build is failing after on of my recent changes, to find out I pushed a commit but the builds are still not working.
That said the tests run fine, it is just TravisCI not working.
I am allowing failure in order to have Selenium tests to run.
Also disabling cache because it could by why the Windows build fails

Signed-off-by: William Desportes <williamdes@wdes.fr>