Fix: Don't add multiple OpenLayers maps, remove listeners on dispose
With 5.1 the map is only created when it is first shown, but every second time
switching to the map a new instance is created because zero is returned when no new
map was created.

OpenLayers adds a resize listener to the window when a target is defined to
handle size adjustments, but this means setTarget(null) has to be called to
fully dispose of the added listener when the map is no longer needed.

Signed-off-by: Maximilian Krög <maxi_kroeg@web.de>