Prevent overriding user setting while auto fixing pma storage
Once the table was created it was overriding the setting. It should respect that the user said the table will be named pma__userconfig_table and not pma__userconfig (the default value).
If the database had the default tables in it, the detection would silently mark them as the right tables totally ignoring the user defined settings for the table names.

Signed-off-by: William Desportes <williamdes@wdes.fr>