Fix #16906 - Fix the caching effect on the feature list
Before this fix you needed to click twice on create tables to see a full list of green checks.
As part of the values where fetched by a global and the feature list values by this $cfgRelation passed as a parameter.

This was visible when the pmadb existed but had no tables in it.

Signed-off-by: William Desportes <williamdes@wdes.fr>