Remove getUnsortedSqlAndSortByKeyDropDown()
This was stupid abstraction. This method did 2 things. When the code that prepares unsorted SQL is moved out of this method, the rest can be absorbed by getSortByKeyDropDown(). This method didn't declare any local variables, it took parameters that were just passed along to other methods. The code to get unsorted SQL was just a single instruction that was self-explanatory. Indexes relied on properties so they didn't need to be a parameter to the other method.

This method only added complexity instead of reducing it.

Signed-off-by: Kamil Tekiela <tekiela246@gmail.com>