Fixed broken filtering - ticket #17593
Fixes #17593
Closes #17610

In (Page-related settings) -> (Where to show the table row links) there
are 4 options: Left, Right, Both, and Nowhere.
The 3 dummy elements (mentioned in this ticket) are only needed for
Left, Both, and Nowhere... and they are NOT needed for "Right".
Since the main thing that distinguishes the "Right" position from the
other 3 is - "colspan", my code checks for the presence of "colspan"
on the first existing (non-dummy) element, and then decides whether
or not to add the 3 additional dummy- elements.

Signed-off-by: Angelo Grebenarov <grebenarov@gmail.com>
Signed-off-by: Maurício Meneghini Fauth <mauricio@fauth.dev>