Create a shared.js file for common JS dependencies
Instead of using the Webpack's dependOn setting, uses the splitChunks
setting to automatically generate the shared.js file.

As the contents of these shared dependencies are being moved to the
shared.js file, the generated file for each dependency is now empty.

Is probably safe to not generate them, but some adjustments are needed
for that first. For example, the AJAX's onload and teardown functions
expects that some of these files exists.

For now, only dependencies that are loaded for all pages should be added
in this shared.js file.

This will facilitate the refactoring of the JS files, but the trade off,
at least for now, is that will be a single large file instead of a few
medium ones.

Signed-off-by: Maurício Meneghini Fauth <mauricio@fauth.dev>