Export plugins refactoring (#18071)
This is a simplification of the code to remove confusing ($row[$j] == '0' || $row[$j] != ''). To make the code simpler, I added a @var annotation and replace for with foreach.

I realize this adds Psalm warnings, and they are valid warnings. But they are because of dead code related to $GLOBALS['what']. This just looks like another nonsense of its own that needs a little more work to be removed so I kept it out of this PR. For the time being, let's keep them in Psalm baseline.

* Simplify ExportCsv
* Simplify ExportHtmlword
* Simplify ExportTexytext
* Fix broken tests

https://github.com/phpmyadmin/phpmyadmin/pull/18071

---------

Signed-off-by: Kamil Tekiela <tekiela246@gmail.com>