Refactor the ServerRequest::getRoute() method
Extracts the redirection to database and table pages to the
HomeController for an actual redirection instead of an implicit one.

Removes the Application::getRequest() method as the ServerRequest object
should be an immutable object.

Adds a Routing::$route static variable to be used at the places that the
ServerRequest object is not available.

Signed-off-by: Maurício Meneghini Fauth <mauricio@fauth.dev>