Avoid auto-updating 'timestamp' fields to NOW()
### Description

The 'firstTimestamp' variable was being set to true for all fields of type 'timestamp', regardless of their default or extra attributes. This led to unintentional updates of these fields to NOW().

Fixes #18677
Row edit has now() preselected on timestamp columns

Solution:
Explicitly set 'firstTimestamp' to false to prevent auto-updating of 'timestamp' fields that don't have a default of "CURRENT_TIMESTAMP" or an extra of "ON UPDATE CURRENT_TIMESTAMP".

Signed-off-by: Nakul Gupta <nakulgupta1042@gmail.com>