Improve collations support for MariaDB 10.10+
### Description

Some more collations were added and structural changes were done in MariaDB 10.10 as per https://jira.mariadb.org/browse/MDEV-27009 which requires changes in how we get the *full* list of collations if we are using this engine.
Since I would not consider this a *bug* submitting to `master` branch.
Since the change  only affects the kind of query which is run, and not the structure of the collation list, no extra tests are added and not existing tests are modified.

Before submitting pull request, please review the following checklist:

- [+] Make sure you have read our [CONTRIBUTING.md](https://github.com/phpmyadmin/phpmyadmin/blob/master/CONTRIBUTING.md) document.
- [+] Make sure you are making a pull request against the correct branch. For example, for bug fixes in a released version use the corresponding QA branch and for new features use the _master_ branch. If you have a doubt, you can ask as a comment in the bug report or on the mailing list.
- [+] Every commit has proper `Signed-off-by` line as described in our [DCO](https://github.com/phpmyadmin/phpmyadmin/blob/master/DCO). This ensures that the work you're submitting is your own creation.
- [+] Every commit has a descriptive commit message.
- [+] Every commit is needed on its own, if you have just minor fixes to previous commits, you can squash them.
- [+] Any new functionality is covered by tests.

Signed-off-by: Dmitrii Kustov <simbiat@outlook.com>