Remove trigger_error from Template class
This will remove the error:

> Error while working with template cache: Unable to create the cache directory

This does not fix the cache issue, but removes the warning.

The current behavior is to disable the cache when there's an issue with
it. However this has bad performance.

The issue is that sometimes Twig generates a different cache key than
the warmed up one and tries to write a new cache, but that's not the
desired behavior.

Signed-off-by: Maurício Meneghini Fauth <mauricio@fauth.dev>