Add `t` function to Twig\I18nExtension
For translations, a Twig function is more versatile than a Twig tag and
has the benefit of auto escaping.

As the phpmyadmin/twig-i18n-extension package is a fork of the Twig's
I18nExtension, the custom function was added to the phpMyAdmin main
repository, this way the cases not used by phpMyAdmin don't need to be
implemented. For example, only the default domain is used and
MoTranslator is always available.

Signed-off-by: Maurício Meneghini Fauth <mauricio@fauth.dev>