Fix formatted sql in grid should be inline text
The css expects all table cells to have a wrapping span, but
the wrapping was only done when no span existed in the cell.
If the formatting is done before the wrapping there already exist
nested spans and no top-level span is created. In this case the
style that makes the sql displayed inline did not apply.

Signed-off-by: Maximilian Krög <maxi_kroeg@web.de>