Fix #19109 - Do not enforce character encoding when checking roles
The functions `getInformationSchemaDataForCreateRequest()` and `getInformationSchemaDataForGranteeRequest()` generate a query that fails with the error `COLLATION 'utf8mb4_general_ci' is not valid for CHARACTER SET 'utf8mb3'` on some servers that got migrated from MySql 5.7. This lead to admin users not being able to create new users via PhpMyAdmin's UI. Instead, the error message "You do not have the privileges to administrate the users!" is being shown. When not forcing the encoding to `utf8mb4_general_ci` in the query, everything works fine.

Signed-off-by: Daniel Schindler<post@danielschindler.de>