Fix export template update error
When updating an export template, the exportType request parameter is
ignored by the controller, because it's not used by the UPDATE
statement. However the Template value object requires a valid ExportType
enum, causing a ValueError to be thrown.

```
ValueError: "" is not a valid backing value for enum PhpMyAdmin\Plugins\ExportType
```

Signed-off-by: Maurício Meneghini Fauth <mauricio@mfauth.net>